Struct iced_widget::canvas::path::lyon_path::builder::NoAttributes

pub struct NoAttributes<B>where
    B: PathBuilder,{ /* private fields */ }

Available on crate feature canvas only.

A convenience wrapper for PathBuilder without custom attributes.

See the PathBuilder trait.

This simply forwards to an underlying PathBuilder implementation, using no attributes.

Implementations

impl<B> NoAttributes<B>where B: PathBuilder,

pub fn wrap(inner: B) -> NoAttributes<B>

pub fn new() -> NoAttributes<B>where B: Default,

pub fn with_capacity(endpoints: usize, ctrl_points: usize) -> NoAttributes<B>where B: Default,

pub fn begin(&mut self, at: Point2D<f32, UnknownUnit>) -> EndpointId

Starts a new sub-path at a given position.

There must be no sub-path in progress when this method is called. at becomes the current position of the sub-path.

pub fn end(&mut self, close: bool)

Ends the current sub path.

A sub-path must be in progress when this method is called. After this method is called, there is no sub-path in progress until begin is called again.

pub fn close(&mut self)

Closes the current sub path.

Shorthand for builder.end(true).

pub fn line_to(&mut self, to: Point2D<f32, UnknownUnit>) -> EndpointId

Adds a line segment to the current sub-path.

A sub-path must be in progress when this method is called.

pub fn quadratic_bezier_to( &mut self, ctrl: Point2D<f32, UnknownUnit>, to: Point2D<f32, UnknownUnit> ) -> EndpointId

Adds a quadratic bézier curve to the current sub-path.

A sub-path must be in progress when this method is called.

pub fn cubic_bezier_to( &mut self, ctrl1: Point2D<f32, UnknownUnit>, ctrl2: Point2D<f32, UnknownUnit>, to: Point2D<f32, UnknownUnit> ) -> EndpointId

Adds a cubic bézier curve to the current sub-path.

A sub-path must be in progress when this method is called.

pub fn reserve(&mut self, endpoints: usize, ctrl_points: usize)

Hints at the builder that a certain number of endpoints and control points will be added.

The Builder implementation may use this information to pre-allocate memory as an optimization.

pub fn path_event( &mut self, event: Event<Point2D<f32, UnknownUnit>, Point2D<f32, UnknownUnit>> )

Applies the provided path event.

By default this calls one of begin, end, line, quadratic_bezier_segment, or cubic_bezier_segment according to the path event.

The requirements for each method apply to the corresponding event.

pub fn add_polygon(&mut self, polygon: Polygon<'_, Point2D<f32, UnknownUnit>>)

Adds a sub-path from a polygon.

There must be no sub-path in progress when this method is called. No sub-path is in progress after the method is called.

pub fn add_point(&mut self, at: Point2D<f32, UnknownUnit>) -> EndpointId

Adds a sub-path containing a single point.

There must be no sub-path in progress when this method is called. No sub-path is in progress after the method is called.

pub fn add_line_segment( &mut self, line: &LineSegment<f32> ) -> (EndpointId, EndpointId)

Adds a sub-path containing a single line segment.

There must be no sub-path in progress when this method is called. No sub-path is in progress after the method is called.

pub fn add_ellipse( &mut self, center: Point2D<f32, UnknownUnit>, radii: Vector2D<f32, UnknownUnit>, x_rotation: Angle<f32>, winding: Winding )

Adds a sub-path containing an ellipse.

There must be no sub-path in progress when this method is called. No sub-path is in progress after the method is called.

pub fn add_circle( &mut self, center: Point2D<f32, UnknownUnit>, radius: f32, winding: Winding )

Adds a sub-path containing a circle.

There must be no sub-path in progress when this method is called. No sub-path is in progress after the method is called.

pub fn add_rectangle( &mut self, rect: &Box2D<f32, UnknownUnit>, winding: Winding )

Adds a sub-path containing a rectangle.

There must be no sub-path in progress when this method is called. No sub-path is in progress after the method is called.

pub fn add_rounded_rectangle( &mut self, rect: &Box2D<f32, UnknownUnit>, radii: &BorderRadii, winding: Winding )

Adds a sub-path containing a rectangle.

There must be no sub-path in progress when this method is called. No sub-path is in progress after the method is called.

pub fn flattened(self, tolerance: f32) -> NoAttributes<Flattened<B>>

Returns a builder that approximates all curves with sequences of line segments.

pub fn transformed<Transform>( self, transform: Transform ) -> NoAttributes<Transformed<B, Transform>>where Transform: Transformation<f32>,

Returns a builder that applies the given transformation to all positions.

pub fn with_svg(self) -> WithSvg<B>

Returns a builder that support SVG commands.

This must be called before starting to add any sub-path.

pub fn build<P>(self) -> Pwhere B: Build<PathType = P>,

Builds a path object, consuming the builder.

pub fn inner(&self) -> &B

pub fn inner_mut(&mut self) -> &mut B

pub fn into_inner(self) -> B

impl NoAttributes<BuilderImpl>

pub fn extend_from_paths(&mut self, paths: &[PathSlice<'_>])