Type Alias iced_widget::canvas::path::lyon_path::geom::euclid::default::SideOffsets2D

pub type SideOffsets2D<T> = SideOffsets2D<T, UnknownUnit>;

Available on crate feature canvas only.

Aliased Type

struct SideOffsets2D<T> {
    pub top: T,
    pub right: T,
    pub bottom: T,
    pub left: T,
    /* private fields */
}

Fields
top: Tright: Tbottom: Tleft: T
Implementations

impl<T, U> SideOffsets2D<T, U>

pub const fn new(top: T, right: T, bottom: T, left: T) -> SideOffsets2D<T, U>

Constructor taking a scalar for each side.

Sides are specified in top-right-bottom-left order following CSS’s convention.

pub fn from_lengths( top: Length<T, U>, right: Length<T, U>, bottom: Length<T, U>, left: Length<T, U> ) -> SideOffsets2D<T, U>

Constructor taking a typed Length for each side.

Sides are specified in top-right-bottom-left order following CSS’s convention.

pub fn from_vectors_outer( min: Vector2D<T, U>, max: Vector2D<T, U> ) -> SideOffsets2D<T, U>where T: Neg<Output = T>,

Construct side offsets from min and a max vector offsets.

The outer rect of the resulting side offsets is equivalent to translating a rectangle’s upper-left corner with the min vector and translating the bottom-right corner with the max vector.

pub fn from_vectors_inner( min: Vector2D<T, U>, max: Vector2D<T, U> ) -> SideOffsets2D<T, U>where T: Neg<Output = T>,

Construct side offsets from min and a max vector offsets.

The inner rect of the resulting side offsets is equivalent to translating a rectangle’s upper-left corner with the min vector and translating the bottom-right corner with the max vector.

pub fn zero() -> SideOffsets2D<T, U>where T: Zero,

Constructor, setting all sides to zero.

pub fn is_zero(&self) -> boolwhere T: Zero + PartialEq<T>,

Returns true if all side offsets are zero.

pub fn new_all_same(all: T) -> SideOffsets2D<T, U>where T: Copy,

Constructor setting the same value to all sides, taking a scalar value directly.

pub fn from_length_all_same(all: Length<T, U>) -> SideOffsets2D<T, U>where T: Copy,

Constructor setting the same value to all sides, taking a typed Length.

pub fn horizontal(&self) -> Twhere T: Copy + Add<T, Output = T>,

pub fn vertical(&self) -> Twhere T: Copy + Add<T, Output = T>,