Struct iced_widget::canvas::path::lyon_path::geom::euclid::Length

#[repr(C)]
pub struct Length<T, Unit>(pub T, _);

Available on crate feature canvas only.

A one-dimensional distance, with value represented by T and unit of measurement Unit.

T can be any numeric type, for example a primitive type like u64 or f32.

Unit is not used in the representation of a Length value. It is used only at compile time to ensure that a Length stored with one unit is converted explicitly before being used in an expression that requires a different unit. It may be a type without values, such as an empty enum.

You can multiply a Length by a scale::Scale to convert it from one unit to another. See the Scale docs for an example.

Tuple Fields
0: T
Implementations

impl<T, U> Length<T, U>

pub const fn new(x: T) -> Length<T, U>

Associate a value with a unit of measure.

impl<T, U> Length<T, U>where T: Clone,

pub fn get(self) -> T

Unpack the underlying value from the wrapper.

pub fn cast_unit<V>(self) -> Length<T, V>

Cast the unit

pub fn lerp(self, other: Length<T, U>, t: T) -> Length<T, U>where T: One + Sub<T, Output = T> + Mul<T, Output = T> + Add<T, Output = T>,

Linearly interpolate between this length and another length.

Example

use euclid::default::Length;

let from = Length::new(0.0);
let to = Length::new(8.0);

assert_eq!(from.lerp(to, -1.0), Length::new(-8.0));
assert_eq!(from.lerp(to,  0.0), Length::new( 0.0));
assert_eq!(from.lerp(to,  0.5), Length::new( 4.0));
assert_eq!(from.lerp(to,  1.0), Length::new( 8.0));
assert_eq!(from.lerp(to,  2.0), Length::new(16.0));

impl<T, U> Length<T, U>where T: PartialOrd<T>,

pub fn min(self, other: Length<T, U>) -> Length<T, U>

Returns minimum between this length and another length.

pub fn max(self, other: Length<T, U>) -> Length<T, U>

Returns maximum between this length and another length.

impl<T, U> Length<T, U>where T: NumCast + Clone,

pub fn cast<NewT>(self) -> Length<NewT, U>where NewT: NumCast,

Cast from one numeric representation to another, preserving the units.

pub fn try_cast<NewT>(self) -> Option<Length<NewT, U>>where NewT: NumCast,

Fallible cast from one numeric representation to another, preserving the units.