iced_widget::canvas::path::lyon_path::geom

Struct LineSegment

pub struct LineSegment<S> {
    pub from: Point2D<S, UnknownUnit>,
    pub to: Point2D<S, UnknownUnit>,
}

A linear segment.

Fields
from: Point2D<S, UnknownUnit>to: Point2D<S, UnknownUnit>
Implementations

impl<S> LineSegment<S>
where S: Scalar,

pub fn sample(&self, t: S) -> Point2D<S, UnknownUnit>

Sample the segment at t (expecting t between 0 and 1).

pub fn x(&self, t: S) -> S

Sample the x coordinate of the segment at t (expecting t between 0 and 1).

pub fn y(&self, t: S) -> S

Sample the y coordinate of the segment at t (expecting t between 0 and 1).

pub fn from(&self) -> Point2D<S, UnknownUnit>

pub fn to(&self) -> Point2D<S, UnknownUnit>

pub fn solve_t_for_x(&self, x: S) -> S

pub fn solve_t_for_y(&self, y: S) -> S

pub fn solve_y_for_x(&self, x: S) -> S

pub fn solve_x_for_y(&self, y: S) -> S

pub fn flip(&self) -> LineSegment<S>

Returns an inverted version of this segment where the beginning and the end points are swapped.

pub fn split_range(&self, t_range: Range<S>) -> LineSegment<S>

Return the sub-segment inside a given range of t.

This is equivalent splitting at the range’s end points.

pub fn split(&self, t: S) -> (LineSegment<S>, LineSegment<S>)

Split this curve into two sub-segments.

pub fn before_split(&self, t: S) -> LineSegment<S>

Return the segment before the split point.

pub fn after_split(&self, t: S) -> LineSegment<S>

Return the segment after the split point.

pub fn split_at_x(&self, x: S) -> (LineSegment<S>, LineSegment<S>)

pub fn bounding_box(&self) -> Box2D<S, UnknownUnit>

Return the smallest rectangle containing this segment.

pub fn to_vector(&self) -> Vector2D<S, UnknownUnit>

Returns the vector between this segment’s from and to points.

pub fn to_line(&self) -> Line<S>

Returns the line containing this segment.

pub fn length(&self) -> S

Computes the length of this segment.

pub fn square_length(&self) -> S

Computes the squared length of this segment.

pub fn set_length(&mut self, new_length: S)

Changes the segment’s length, moving destination point.

pub fn mid_point(&mut self) -> Point2D<S, UnknownUnit>

Computes third mid-point of this segment.

pub fn translate(&mut self, by: Vector2D<S, UnknownUnit>) -> LineSegment<S>

pub fn transformed<T>(&self, transform: &T) -> LineSegment<S>
where T: Transformation<S>,

Applies the transform to this segment and returns the results.

pub fn intersection_t(&self, other: &LineSegment<S>) -> Option<(S, S)>

Computes the intersection (if any) between this segment and another one.

The result is provided in the form of the t parameter of each segment. To get the intersection point, sample one of the segments at the corresponding value.

pub fn intersection( &self, other: &LineSegment<S>, ) -> Option<Point2D<S, UnknownUnit>>

pub fn line_intersection_t(&self, line: &Line<S>) -> Option<S>

pub fn line_intersection( &self, line: &Line<S>, ) -> Option<Point2D<S, UnknownUnit>>

pub fn horizontal_line_intersection_t(&self, y: S) -> Option<S>

pub fn vertical_line_intersection_t(&self, x: S) -> Option<S>

pub fn horizontal_line_intersection( &self, y: S, ) -> Option<Point2D<S, UnknownUnit>>

pub fn vertical_line_intersection( &self, x: S, ) -> Option<Point2D<S, UnknownUnit>>

pub fn intersects(&self, other: &LineSegment<S>) -> bool

pub fn intersects_line(&self, line: &Line<S>) -> bool

pub fn overlaps_line(&self, line: &Line<S>) -> bool

pub fn overlaps_segment(&self, other: &LineSegment<S>) -> bool

pub fn contains_segment(&self, other: &LineSegment<S>) -> bool

pub fn clipped_x(&self, clip: Range<S>) -> Option<LineSegment<S>>

Horizontally clip this segment against a range of the x axis.

pub fn clipped_y(&self, clip: Range<S>) -> Option<LineSegment<S>>

Vertically clip this segment against a range of the y axis.

pub fn clipped(&self, clip: &Box2D<S, UnknownUnit>) -> Option<LineSegment<S>>

Clip this segment against a rectangle.

pub fn distance_to_point(&self, p: Point2D<S, UnknownUnit>) -> S

Computes the distance between this segment and a point.

pub fn square_distance_to_point(&self, p: Point2D<S, UnknownUnit>) -> S

Computes the squared distance between this segment and a point.

Can be useful to save a square root and a division when comparing against a distance that can be squared.

pub fn closest_point( &self, p: Point2D<S, UnknownUnit>, ) -> Point2D<S, UnknownUnit>

Computes the closest point on this segment to p.

pub fn to_f32(&self) -> LineSegment<f32>

pub fn to_f64(&self) -> LineSegment<f64>