iced_widget::canvas::path::lyon_path::geom

Type Alias Size

pub type Size<T> = Size2D<T, UnknownUnit>;

Aliased Type

struct Size<T> {
    pub width: T,
    pub height: T,
}

Fields
width: T

The extent of the element in the U units along the x axis (usually horizontal).

height: T

The extent of the element in the U units along the y axis (usually vertical).