Type Alias iced_widget::canvas::path::lyon_path::geom::Transform

pub type Transform<S> = Transform2D<S, UnknownUnit, UnknownUnit>;

Available on crate feature canvas only.

Alias for euclid::default::Transform2D

Aliased Type

struct Transform<S> {
    pub m11: S,
    pub m12: S,
    pub m21: S,
    pub m22: S,
    pub m31: S,
    pub m32: S,
    /* private fields */
}

Fields
m11: Sm12: Sm21: Sm22: Sm31: Sm32: S
Implementations

impl<T, Src, Dst> Transform2D<T, Src, Dst>

pub const fn new( m11: T, m12: T, m21: T, m22: T, m31: T, m32: T ) -> Transform2D<T, Src, Dst>

Create a transform specifying its components in using the column-major-column-vector matrix notation.

For example, the translation terms m31 and m32 are the last two parameters parameters.

use euclid::default::Transform2D;
let tx = 1.0;
let ty = 2.0;
let translation = Transform2D::new(
  1.0, 0.0,
  0.0, 1.0,
  tx,  ty,
);

pub fn approx_eq(&self, other: &Transform2D<T, Src, Dst>) -> boolwhere T: ApproxEq<T>,

Returns true is this transform is approximately equal to the other one, using T’s default epsilon value.

The same as ApproxEq::approx_eq() but available without importing trait.

pub fn approx_eq_eps(&self, other: &Transform2D<T, Src, Dst>, eps: &T) -> boolwhere T: ApproxEq<T>,

Returns true is this transform is approximately equal to the other one, using a provided epsilon value.

The same as ApproxEq::approx_eq_eps() but available without importing trait.

impl<T, Src, Dst> Transform2D<T, Src, Dst>where T: Copy,

pub fn to_array(&self) -> [T; 6]

Returns an array containing this transform’s terms.

The terms are laid out in the same order as they are specified in Transform2D::new, that is following the column-major-column-vector matrix notation.

For example the translation terms are found in the last two slots of the array.

pub fn to_array_transposed(&self) -> [T; 6]

Returns an array containing this transform’s terms transposed.

The terms are laid out in transposed order from the same order of Transform3D::new and Transform3D::to_array, that is following the row-major-column-vector matrix notation.

For example the translation terms are found at indices 2 and 5 in the array.

pub fn to_arrays(&self) -> [[T; 2]; 3]

Equivalent to to_array with elements packed two at a time in an array of arrays.

pub fn from_array(array: [T; 6]) -> Transform2D<T, Src, Dst>

Create a transform providing its components via an array of 6 elements instead of as individual parameters.

The order of the components corresponds to the column-major-column-vector matrix notation (the same order as Transform2D::new).

pub fn from_arrays(array: [[T; 2]; 3]) -> Transform2D<T, Src, Dst>

Equivalent to from_array with elements packed two at a time in an array of arrays.

The order of the components corresponds to the column-major-column-vector matrix notation (the same order as Transform3D::new).

pub fn to_untyped(&self) -> Transform2D<T, UnknownUnit, UnknownUnit>

Drop the units, preserving only the numeric value.

pub fn from_untyped( p: &Transform2D<T, UnknownUnit, UnknownUnit> ) -> Transform2D<T, Src, Dst>

Tag a unitless value with units.

pub fn with_source<NewSrc>(&self) -> Transform2D<T, NewSrc, Dst>

Returns the same transform with a different source unit.

pub fn with_destination<NewDst>(&self) -> Transform2D<T, Src, NewDst>

Returns the same transform with a different destination unit.

pub fn to_3d(&self) -> Transform3D<T, Src, Dst>where T: Zero + One,

Create a 3D transform from the current transform

impl<T, Src, Dst> Transform2D<T, Src, Dst>where T: NumCast + Copy,

pub fn cast<NewT>(&self) -> Transform2D<NewT, Src, Dst>where NewT: NumCast,

Cast from one numeric representation to another, preserving the units.

pub fn try_cast<NewT>(&self) -> Option<Transform2D<NewT, Src, Dst>>where NewT: NumCast,

Fallible cast from one numeric representation to another, preserving the units.

impl<T, Src, Dst> Transform2D<T, Src, Dst>where T: Zero + One,

pub fn identity() -> Transform2D<T, Src, Dst>

Create an identity matrix:

1 0
0 1
0 0

impl<T, Src, Dst> Transform2D<T, Src, Dst>where T: Copy + Add<T, Output = T> + Mul<T, Output = T>,

Methods for combining generic transformations

pub fn then<NewDst>( &self, mat: &Transform2D<T, Dst, NewDst> ) -> Transform2D<T, Src, NewDst>

Returns the multiplication of the two matrices such that mat’s transformation applies after self’s transformation.

impl<T, Src, Dst> Transform2D<T, Src, Dst>where T: Zero + One,

Methods for creating and combining translation transformations

pub fn translation(x: T, y: T) -> Transform2D<T, Src, Dst>

Create a 2d translation transform:

1 0
0 1
x y

pub fn then_translate(&self, v: Vector2D<T, Dst>) -> Transform2D<T, Src, Dst>where T: Copy + Add<T, Output = T> + Mul<T, Output = T>,

Applies a translation after self’s transformation and returns the resulting transform.

pub fn pre_translate(&self, v: Vector2D<T, Src>) -> Transform2D<T, Src, Dst>where T: Copy + Add<T, Output = T> + Mul<T, Output = T>,

Applies a translation before self’s transformation and returns the resulting transform.

impl<T, Src, Dst> Transform2D<T, Src, Dst>where T: Copy + Add<T, Output = T> + Sub<T, Output = T> + Mul<T, Output = T> + Zero + Trig,

Methods for creating and combining rotation transformations

pub fn rotation(theta: Angle<T>) -> Transform2D<T, Src, Dst>

Returns a rotation transform.

pub fn then_rotate(&self, theta: Angle<T>) -> Transform2D<T, Src, Dst>

Applies a rotation after self’s transformation and returns the resulting transform.

pub fn pre_rotate(&self, theta: Angle<T>) -> Transform2D<T, Src, Dst>

Applies a rotation before self’s transformation and returns the resulting transform.

impl<T, Src, Dst> Transform2D<T, Src, Dst>

Methods for creating and combining scale transformations

pub fn scale(x: T, y: T) -> Transform2D<T, Src, Dst>where T: Zero,

Create a 2d scale transform:

x 0
0 y
0 0

pub fn then_scale(&self, x: T, y: T) -> Transform2D<T, Src, Dst>where T: Copy + Add<T, Output = T> + Mul<T, Output = T> + Zero,

Applies a scale after self’s transformation and returns the resulting transform.

pub fn pre_scale(&self, x: T, y: T) -> Transform2D<T, Src, Dst>where T: Copy + Mul<T, Output = T>,

Applies a scale before self’s transformation and returns the resulting transform.

impl<T, Src, Dst> Transform2D<T, Src, Dst>where T: Copy + Add<T, Output = T> + Mul<T, Output = T>,

Methods for apply transformations to objects

pub fn transform_point(&self, point: Point2D<T, Src>) -> Point2D<T, Dst>

Returns the given point transformed by this transform.

pub fn transform_vector(&self, vec: Vector2D<T, Src>) -> Vector2D<T, Dst>

Returns the given vector transformed by this matrix.

pub fn outer_transformed_rect(&self, rect: &Rect<T, Src>) -> Rect<T, Dst>where T: Sub<T, Output = T> + Zero + PartialOrd<T>,

Returns a rectangle that encompasses the result of transforming the given rectangle by this transform.

pub fn outer_transformed_box(&self, b: &Box2D<T, Src>) -> Box2D<T, Dst>where T: Sub<T, Output = T> + Zero + PartialOrd<T>,

Returns a box that encompasses the result of transforming the given box by this transform.

impl<T, Src, Dst> Transform2D<T, Src, Dst>where T: Copy + Sub<T, Output = T> + Mul<T, Output = T> + Div<T, Output = T> + PartialEq<T> + Zero + One,

pub fn determinant(&self) -> T

Computes and returns the determinant of this transform.

pub fn is_invertible(&self) -> bool

Returns whether it is possible to compute the inverse transform.

pub fn inverse(&self) -> Option<Transform2D<T, Dst, Src>>

Returns the inverse transform if possible.