iced_widget::combo_box

Struct ComboBox

pub struct ComboBox<'a, T, Message, Theme = Theme, Renderer = Renderer>
where
    Theme: Catalog,
    Renderer: Renderer,
{ /* private fields */ }

A widget for searching and selecting a single value from a list of options.

Example

use iced::widget::combo_box;

struct State {
   fruits: combo_box::State<Fruit>,
   favorite: Option<Fruit>,
}

#[derive(Debug, Clone)]
enum Fruit {
    Apple,
    Orange,
    Strawberry,
    Tomato,
}

#[derive(Debug, Clone)]
enum Message {
    FruitSelected(Fruit),
}

fn view(state: &State) -> Element<'_, Message> {
    combo_box(
        &state.fruits,
        "Select your favorite fruit...",
        state.favorite.as_ref(),
        Message::FruitSelected
    )
    .into()
}

fn update(state: &mut State, message: Message) {
    match message {
        Message::FruitSelected(fruit) => {
            state.favorite = Some(fruit);
        }
    }
}

impl std::fmt::Display for Fruit {
    fn fmt(&self, f: &mut std::fmt::Formatter<'_>) -> std::fmt::Result {
        f.write_str(match self {
            Self::Apple => "Apple",
            Self::Orange => "Orange",
            Self::Strawberry => "Strawberry",
            Self::Tomato => "Tomato",
        })
    }
}

Implementations

impl<'a, T, Message, Theme, Renderer> ComboBox<'a, T, Message, Theme, Renderer>
where T: Display + Clone, Theme: Catalog, Renderer: Renderer,

pub fn new( state: &'a State<T>, placeholder: &str, selection: Option<&T>, on_selected: impl Fn(T) -> Message + 'static, ) -> Self

Creates a new ComboBox with the given list of options, a placeholder, the current selected value, and the message to produce when an option is selected.

pub fn on_input(self, on_input: impl Fn(String) -> Message + 'static) -> Self

Sets the message that should be produced when some text is typed into the TextInput of the ComboBox.

pub fn on_option_hovered( self, on_option_hovered: impl Fn(T) -> Message + 'static, ) -> Self

Sets the message that will be produced when an option of the ComboBox is hovered using the arrow keys.

pub fn on_open(self, message: Message) -> Self

Sets the message that will be produced when the ComboBox is opened.

pub fn on_close(self, message: Message) -> Self

Sets the message that will be produced when the outside area of the ComboBox is pressed.

pub fn padding(self, padding: impl Into<Padding>) -> Self

Sets the [Padding] of the ComboBox.

pub fn font(self, font: Renderer::Font) -> Self

Sets the Renderer::Font of the ComboBox.

pub fn icon(self, icon: Icon<Renderer::Font>) -> Self

Sets the text_input::Icon of the ComboBox.

pub fn size(self, size: impl Into<Pixels>) -> Self

Sets the text sixe of the ComboBox.

pub fn line_height(self, line_height: impl Into<LineHeight>) -> Self

Sets the LineHeight of the ComboBox.

pub fn width(self, width: impl Into<Length>) -> Self

Sets the width of the ComboBox.

pub fn input_style(self, style: impl Fn(&Theme, Status) -> Style + 'a) -> Self
where <Theme as Catalog>::Class<'a>: From<StyleFn<'a, Theme>>,

Sets the style of the input of the ComboBox.

pub fn menu_style(self, style: impl Fn(&Theme) -> Style + 'a) -> Self
where <Theme as Catalog>::Class<'a>: From<StyleFn<'a, Theme>>,

Sets the style of the menu of the ComboBox.

pub fn input_class( self, class: impl Into<<Theme as Catalog>::Class<'a>>, ) -> Self

Available on crate feature advanced only.

Sets the style class of the input of the ComboBox.

pub fn menu_class(self, class: impl Into<<Theme as Catalog>::Class<'a>>) -> Self

Available on crate feature advanced only.

Sets the style class of the menu of the ComboBox.