iced_widget::combo_box

Trait Catalog

pub trait Catalog: Catalog + Catalog {
    // Provided methods
    fn default_input<'a>() -> <Self as Catalog>::Class<'a> { ... }
    fn default_menu<'a>() -> <Self as Catalog>::Class<'a> { ... }
}

The theme catalog of a ComboBox.

Provided Methods

fn default_input<'a>() -> <Self as Catalog>::Class<'a>

The default class for the text input of the ComboBox.

fn default_menu<'a>() -> <Self as Catalog>::Class<'a>

The default class for the menu of the ComboBox.

Dyn Compatibility

This trait is not dyn compatible.

In older versions of Rust, dyn compatibility was called "object safety", so this trait is not object safe.