iced_widget::keyed::column

Struct Column

pub struct Column<'a, Key, Message, Theme = Theme, Renderer = Renderer>
where
    Key: Copy + PartialEq,
{ /* private fields */ }

A container that distributes its contents vertically while keeping continuity.

Example

use iced::widget::{keyed_column, text};

enum Message {
    // ...
}

fn view(state: &State) -> Element<'_, Message> {
    keyed_column((0..=100).map(|i| {
        (i, text!("Item {i}").into())
    })).into()
}

Implementations

impl<'a, Key, Message, Theme, Renderer> Column<'a, Key, Message, Theme, Renderer>
where Key: Copy + PartialEq, Renderer: Renderer,

pub fn new() -> Self

Creates an empty Column.

pub fn from_vecs( keys: Vec<Key>, children: Vec<Element<'a, Message, Theme, Renderer>>, ) -> Self

Creates a Column from already allocated Vecs.

Keep in mind that the Column will not inspect the Vecs, which means it won’t automatically adapt to the sizing strategy of its contents.

If any of the children have a [Length::Fill] strategy, you will need to call Column::width or Column::height accordingly.

pub fn with_capacity(capacity: usize) -> Self

Creates a Column with the given capacity.

pub fn with_children( children: impl IntoIterator<Item = (Key, Element<'a, Message, Theme, Renderer>)>, ) -> Self

Creates a Column with the given elements.

pub fn spacing(self, amount: impl Into<Pixels>) -> Self

Sets the vertical spacing between elements.

Custom margins per element do not exist in iced. You should use this method instead! While less flexible, it helps you keep spacing between elements consistent.

pub fn padding<P: Into<Padding>>(self, padding: P) -> Self

Sets the [Padding] of the Column.

pub fn width(self, width: impl Into<Length>) -> Self

Sets the width of the Column.

pub fn height(self, height: impl Into<Length>) -> Self

Sets the height of the Column.

pub fn max_width(self, max_width: impl Into<Pixels>) -> Self

Sets the maximum width of the Column.

pub fn align_items(self, align: Alignment) -> Self

Sets the horizontal alignment of the contents of the Column .

pub fn push( self, key: Key, child: impl Into<Element<'a, Message, Theme, Renderer>>, ) -> Self

Adds an element to the Column.

pub fn push_maybe( self, key: Key, child: Option<impl Into<Element<'a, Message, Theme, Renderer>>>, ) -> Self

Adds an element to the Column, if Some.

pub fn extend( self, children: impl IntoIterator<Item = (Key, Element<'a, Message, Theme, Renderer>)>, ) -> Self

Extends the Column with the given children.