Trait iced_widget::pane_grid::StyleSheet

pub trait StyleSheet {
    type Style: Default;

    // Required methods
    fn hovered_region(&self, style: &Self::Style) -> Appearance;
    fn picked_split(&self, style: &Self::Style) -> Option<Line>;
    fn hovered_split(&self, style: &Self::Style) -> Option<Line>;
}

A set of rules that dictate the style of a container.

Required Associated Types

type Style: Default

The supported style of the StyleSheet.

Required Methods

fn hovered_region(&self, style: &Self::Style) -> Appearance

The Appearance to draw when a pane is hovered.

fn picked_split(&self, style: &Self::Style) -> Option<Line>

The Line to draw when a split is picked.

fn hovered_split(&self, style: &Self::Style) -> Option<Line>

The Line to draw when a split is hovered.