iced_widget::pick_list

Struct PickList

pub struct PickList<'a, T, L, V, Message, Theme = Theme, Renderer = Renderer>
where
    T: ToString + PartialEq + Clone,
    L: Borrow<[T]> + 'a,
    V: Borrow<T> + 'a,
    Theme: Catalog,
    Renderer: Renderer,
{ /* private fields */ }

A widget for selecting a single value from a list of options.

Example

use iced::widget::pick_list;

struct State {
   favorite: Option<Fruit>,
}

#[derive(Debug, Clone, Copy, PartialEq, Eq)]
enum Fruit {
    Apple,
    Orange,
    Strawberry,
    Tomato,
}

#[derive(Debug, Clone)]
enum Message {
    FruitSelected(Fruit),
}

fn view(state: &State) -> Element<'_, Message> {
    let fruits = [
        Fruit::Apple,
        Fruit::Orange,
        Fruit::Strawberry,
        Fruit::Tomato,
    ];

    pick_list(
        fruits,
        state.favorite,
        Message::FruitSelected,
    )
    .placeholder("Select your favorite fruit...")
    .into()
}

fn update(state: &mut State, message: Message) {
    match message {
        Message::FruitSelected(fruit) => {
            state.favorite = Some(fruit);
        }
    }
}

impl std::fmt::Display for Fruit {
    fn fmt(&self, f: &mut std::fmt::Formatter<'_>) -> std::fmt::Result {
        f.write_str(match self {
            Self::Apple => "Apple",
            Self::Orange => "Orange",
            Self::Strawberry => "Strawberry",
            Self::Tomato => "Tomato",
        })
    }
}

Implementations

impl<'a, T, L, V, Message, Theme, Renderer> PickList<'a, T, L, V, Message, Theme, Renderer>
where T: ToString + PartialEq + Clone, L: Borrow<[T]> + 'a, V: Borrow<T> + 'a, Message: Clone, Theme: Catalog, Renderer: Renderer,

pub fn new( options: L, selected: Option<V>, on_select: impl Fn(T) -> Message + 'a, ) -> Self

Creates a new PickList with the given list of options, the current selected value, and the message to produce when an option is selected.

pub fn placeholder(self, placeholder: impl Into<String>) -> Self

Sets the placeholder of the PickList.

pub fn width(self, width: impl Into<Length>) -> Self

Sets the width of the PickList.

pub fn padding<P: Into<Padding>>(self, padding: P) -> Self

Sets the [Padding] of the PickList.

pub fn text_size(self, size: impl Into<Pixels>) -> Self

Sets the text size of the PickList.

pub fn text_line_height(self, line_height: impl Into<LineHeight>) -> Self

Sets the text text::LineHeight of the PickList.

pub fn text_shaping(self, shaping: Shaping) -> Self

Sets the text::Shaping strategy of the PickList.

pub fn font(self, font: impl Into<Renderer::Font>) -> Self

Sets the font of the PickList.

pub fn handle(self, handle: Handle<Renderer::Font>) -> Self

Sets the Handle of the PickList.

pub fn on_open(self, on_open: Message) -> Self

Sets the message that will be produced when the PickList is opened.

pub fn on_close(self, on_close: Message) -> Self

Sets the message that will be produced when the PickList is closed.

pub fn style(self, style: impl Fn(&Theme, Status) -> Style + 'a) -> Self
where <Theme as Catalog>::Class<'a>: From<StyleFn<'a, Theme>>,

Sets the style of the PickList.

pub fn menu_style(self, style: impl Fn(&Theme) -> Style + 'a) -> Self
where <Theme as Catalog>::Class<'a>: From<StyleFn<'a, Theme>>,

Sets the style of the Menu.

pub fn class(self, class: impl Into<<Theme as Catalog>::Class<'a>>) -> Self

Available on crate feature advanced only.

Sets the style class of the PickList.

pub fn menu_class(self, class: impl Into<<Theme as Catalog>::Class<'a>>) -> Self

Available on crate feature advanced only.

Sets the style class of the Menu.