Module iced_widget::rule

Display a horizontal or vertical rule for dividing content.

Structs

Rule

Display a horizontal or vertical rule for dividing content.

Style

The appearance of a rule.

Enums

FillMode

The fill mode of a rule.

Traits

Catalog

The theme catalog of a Rule.

Functions

default

The default styling of a Rule.

Type Aliases

StyleFn

A styling function for a Rule.