iced_widget::scrollable

Struct Scrollable

pub struct Scrollable<'a, Message, Theme = Theme, Renderer = Renderer>
where
    Theme: Catalog,
    Renderer: Renderer,
{ /* private fields */ }

A widget that can vertically display an infinite amount of content with a scrollbar.

Example

use iced::widget::{column, scrollable, vertical_space};

enum Message {
    // ...
}

fn view(state: &State) -> Element<'_, Message> {
    scrollable(column![
        "Scroll me!",
        vertical_space().height(3000),
        "You did it!",
    ]).into()
}

Implementations

impl<'a, Message, Theme, Renderer> Scrollable<'a, Message, Theme, Renderer>
where Theme: Catalog, Renderer: Renderer,

pub fn new(content: impl Into<Element<'a, Message, Theme, Renderer>>) -> Self

Creates a new vertical Scrollable.

pub fn with_direction( content: impl Into<Element<'a, Message, Theme, Renderer>>, direction: impl Into<Direction>, ) -> Self

Creates a new Scrollable with the given Direction.

pub fn horizontal(self) -> Self

Makes the Scrollable scroll horizontally, with default Scrollbar settings.

pub fn direction(self, direction: impl Into<Direction>) -> Self

Sets the Direction of the Scrollable.

pub fn id(self, id: impl Into<Id>) -> Self

Sets the Id of the Scrollable.

pub fn width(self, width: impl Into<Length>) -> Self

Sets the width of the Scrollable.

pub fn height(self, height: impl Into<Length>) -> Self

Sets the height of the Scrollable.

pub fn on_scroll(self, f: impl Fn(Viewport) -> Message + 'a) -> Self

Sets a function to call when the Scrollable is scrolled.

The function takes the Viewport of the Scrollable

pub fn anchor_top(self) -> Self

Anchors the vertical Scrollable direction to the top.

pub fn anchor_bottom(self) -> Self

Anchors the vertical Scrollable direction to the bottom.

pub fn anchor_left(self) -> Self

Anchors the horizontal Scrollable direction to the left.

pub fn anchor_right(self) -> Self

Anchors the horizontal Scrollable direction to the right.

pub fn anchor_x(self, alignment: Anchor) -> Self

Sets the Anchor of the horizontal direction of the Scrollable, if applicable.

pub fn anchor_y(self, alignment: Anchor) -> Self

Sets the Anchor of the vertical direction of the Scrollable, if applicable.

pub fn spacing(self, new_spacing: impl Into<Pixels>) -> Self

Embeds the Scrollbar into the Scrollable, instead of floating on top of the content.

The spacing provided will be used as space between the Scrollbar and the contents of the Scrollable.

pub fn style(self, style: impl Fn(&Theme, Status) -> Style + 'a) -> Self
where Theme::Class<'a>: From<StyleFn<'a, Theme>>,

Sets the style of this Scrollable.

pub fn class(self, class: impl Into<Theme::Class<'a>>) -> Self

Available on crate feature advanced only.

Sets the style class of the Scrollable.