Trait iced_widget::scrollable::StyleSheet

pub trait StyleSheet {
    type Style: Default;

    // Required methods
    fn active(&self, style: &Self::Style) -> Appearance;
    fn hovered(
        &self,
        style: &Self::Style,
        is_mouse_over_scrollbar: bool
    ) -> Appearance;

    // Provided method
    fn dragging(&self, style: &Self::Style) -> Appearance { ... }
}

A set of rules that dictate the style of a scrollable.

Required Associated Types

type Style: Default

The supported style of the StyleSheet.

Required Methods

fn active(&self, style: &Self::Style) -> Appearance

Produces the Appearance of an active scrollable.

fn hovered( &self, style: &Self::Style, is_mouse_over_scrollbar: bool ) -> Appearance

Produces the Appearance of a scrollable when it is being hovered.

Provided Methods

fn dragging(&self, style: &Self::Style) -> Appearance

Produces the Appearance of a scrollable when it is being dragged.