Trait iced_widget::shader::Program

pub trait Program<Message> {
    type State: Default + 'static;
    type Primitive: Primitive + 'static;

    // Required method
    fn draw(
        &self,
        state: &Self::State,
        cursor: Cursor,
        bounds: Rectangle
    ) -> Self::Primitive;

    // Provided methods
    fn update(
        &self,
        _state: &mut Self::State,
        _event: Event,
        _bounds: Rectangle,
        _cursor: Cursor,
        _shell: &mut Shell<'_, Message>
    ) -> (Status, Option<Message>) { ... }
    fn mouse_interaction(
        &self,
        _state: &Self::State,
        _bounds: Rectangle,
        _cursor: Cursor
    ) -> Interaction { ... }
}

Available on crate feature wgpu only.

The state and logic of a Shader widget.

A Program can mutate the internal state of a Shader widget and produce messages for an application.

Required Associated Types

type State: Default + 'static

The internal state of the Program.

type Primitive: Primitive + 'static

The type of primitive this Program can draw.

Required Methods

fn draw( &self, state: &Self::State, cursor: Cursor, bounds: Rectangle ) -> Self::Primitive

Draws the Primitive.

Provided Methods

fn update( &self, _state: &mut Self::State, _event: Event, _bounds: Rectangle, _cursor: Cursor, _shell: &mut Shell<'_, Message> ) -> (Status, Option<Message>)

Update the internal State of the Program. This can be used to reflect state changes based on mouse & other events. You can use the [Shell] to publish messages, request a redraw for the window, etc.

By default, this method does and returns nothing.

fn mouse_interaction( &self, _state: &Self::State, _bounds: Rectangle, _cursor: Cursor ) -> Interaction

Returns the current mouse interaction of the Program.

The interaction returned will be in effect even if the cursor position is out of bounds of the Shader’s program.