iced_widget

Struct Responsive

pub struct Responsive<'a, Message, Theme = Theme, Renderer = Renderer> { /* private fields */ }

Available on crate feature lazy only.

A widget that is aware of its dimensions.

A Responsive widget will always try to fill all the available space of its parent.

Implementations

impl<'a, Message, Theme, Renderer> Responsive<'a, Message, Theme, Renderer>
where Renderer: Renderer,

pub fn new( view: impl Fn(Size) -> Element<'a, Message, Theme, Renderer> + 'a, ) -> Self

Creates a new Responsive widget with a closure that produces its contents.

The view closure will be provided with the current Size of the Responsive widget and, therefore, can be used to build the contents of the widget in a responsive way.