iced_widget::text

Struct Rich

pub struct Rich<'a, Link, Message, Theme = Theme, Renderer = Renderer>
where
    Link: Clone + 'static,
    Theme: Catalog,
    Renderer: Renderer,
{ /* private fields */ }

A bunch of Rich text.

Implementations

impl<'a, Link, Message, Theme, Renderer> Rich<'a, Link, Message, Theme, Renderer>
where Link: Clone + 'static, Theme: Catalog, Renderer: Renderer, Renderer::Font: 'a,

pub fn new() -> Self

Creates a new empty Rich text.

pub fn with_spans( spans: impl AsRef<[Span<'a, Link, Renderer::Font>]> + 'a, ) -> Self

Creates a new Rich text with the given text spans.

pub fn size(self, size: impl Into<Pixels>) -> Self

Sets the default size of the Rich text.

pub fn line_height(self, line_height: impl Into<LineHeight>) -> Self

Sets the default LineHeight of the Rich text.

pub fn font(self, font: impl Into<Renderer::Font>) -> Self

Sets the default font of the Rich text.

pub fn width(self, width: impl Into<Length>) -> Self

Sets the width of the Rich text boundaries.

pub fn height(self, height: impl Into<Length>) -> Self

Sets the height of the Rich text boundaries.

pub fn center(self) -> Self

Centers the Rich text, both horizontally and vertically.

pub fn align_x(self, alignment: impl Into<Alignment>) -> Self

Sets the [alignment::Horizontal] of the Rich text.

pub fn align_y(self, alignment: impl Into<Vertical>) -> Self

Sets the [alignment::Vertical] of the Rich text.

pub fn wrapping(self, wrapping: Wrapping) -> Self

Sets the Wrapping strategy of the Rich text.

pub fn on_link_click(self, on_link_click: impl Fn(Link) -> Message + 'a) -> Self

Sets the message that will be produced when a link of the Rich text is clicked.

If the spans of the Rich text contain no links, you may need to call this method with on_link_click(never) in order for the compiler to infer the proper Link generic type.

pub fn style(self, style: impl Fn(&Theme) -> Style + 'a) -> Self
where Theme::Class<'a>: From<StyleFn<'a, Theme>>,

Sets the default style of the Rich text.

pub fn color(self, color: impl Into<Color>) -> Self
where Theme::Class<'a>: From<StyleFn<'a, Theme>>,

Sets the default [Color] of the Rich text.

pub fn color_maybe(self, color: Option<impl Into<Color>>) -> Self
where Theme::Class<'a>: From<StyleFn<'a, Theme>>,

Sets the default [Color] of the Rich text, if Some.

pub fn class(self, class: impl Into<Theme::Class<'a>>) -> Self

Available on crate feature advanced only.

Sets the default style class of the Rich text.