iced_widget::text_editor

Struct TextEditor

pub struct TextEditor<'a, Highlighter, Message, Theme = Theme, Renderer = Renderer>
where
    Highlighter: Highlighter,
    Theme: Catalog,
    Renderer: Renderer,
{ /* private fields */ }

A multi-line text input.

Example

use iced::widget::text_editor;

struct State {
   content: text_editor::Content,
}

#[derive(Debug, Clone)]
enum Message {
    Edit(text_editor::Action)
}

fn view(state: &State) -> Element<'_, Message> {
    text_editor(&state.content)
        .placeholder("Type something here...")
        .on_action(Message::Edit)
        .into()
}

fn update(state: &mut State, message: Message) {
    match message {
        Message::Edit(action) => {
            state.content.perform(action);
        }
    }
}

Implementations

impl<'a, Message, Theme, Renderer> TextEditor<'a, PlainText, Message, Theme, Renderer>
where Theme: Catalog, Renderer: Renderer,

pub fn new(content: &'a Content<Renderer>) -> Self

Creates new TextEditor with the given Content.

impl<'a, Highlighter, Message, Theme, Renderer> TextEditor<'a, Highlighter, Message, Theme, Renderer>
where Highlighter: Highlighter, Theme: Catalog, Renderer: Renderer,

pub fn placeholder(self, placeholder: impl IntoFragment<'a>) -> Self

Sets the placeholder of the TextEditor.

pub fn width(self, width: impl Into<Pixels>) -> Self

Sets the width of the TextEditor.

pub fn height(self, height: impl Into<Length>) -> Self

Sets the height of the TextEditor.

pub fn min_height(self, min_height: impl Into<Pixels>) -> Self

Sets the minimum height of the TextEditor.

pub fn max_height(self, max_height: impl Into<Pixels>) -> Self

Sets the maximum height of the TextEditor.

pub fn on_action(self, on_edit: impl Fn(Action) -> Message + 'a) -> Self

Sets the message that should be produced when some action is performed in the TextEditor.

If this method is not called, the TextEditor will be disabled.

pub fn font(self, font: impl Into<Renderer::Font>) -> Self

Sets the Font of the TextEditor.

pub fn size(self, size: impl Into<Pixels>) -> Self

Sets the text size of the TextEditor.

pub fn line_height(self, line_height: impl Into<LineHeight>) -> Self

Sets the text::LineHeight of the TextEditor.

pub fn padding(self, padding: impl Into<Padding>) -> Self

Sets the [Padding] of the TextEditor.

pub fn wrapping(self, wrapping: Wrapping) -> Self

Sets the Wrapping strategy of the TextEditor.

pub fn highlight( self, syntax: &str, theme: Theme, ) -> TextEditor<'a, Highlighter, Message, Theme, Renderer>
where Renderer: Renderer<Font = Font>,

Available on crate feature highlighter only.

Highlights the TextEditor using the given syntax and theme.

pub fn highlight_with<H: Highlighter>( self, settings: H::Settings, to_format: fn(&H::Highlight, &Theme) -> Format<Renderer::Font>, ) -> TextEditor<'a, H, Message, Theme, Renderer>

Highlights the TextEditor with the given Highlighter and a strategy to turn its highlights into some text format.

pub fn key_binding( self, key_binding: impl Fn(KeyPress) -> Option<Binding<Message>> + 'a, ) -> Self

Sets the closure to produce key bindings on key presses.

See Binding for the list of available bindings.

pub fn style(self, style: impl Fn(&Theme, Status) -> Style + 'a) -> Self
where Theme::Class<'a>: From<StyleFn<'a, Theme>>,

Sets the style of the TextEditor.

pub fn class(self, class: impl Into<Theme::Class<'a>>) -> Self

Available on crate feature advanced only.

Sets the style class of the TextEditor.