iced_widget::text_input

Struct TextInput

pub struct TextInput<'a, Message, Theme = Theme, Renderer = Renderer>
where
    Theme: Catalog,
    Renderer: Renderer,
{ /* private fields */ }

A field that can be filled with text.

Example

use iced::widget::text_input;

struct State {
   content: String,
}

#[derive(Debug, Clone)]
enum Message {
    ContentChanged(String)
}

fn view(state: &State) -> Element<'_, Message> {
    text_input("Type something here...", &state.content)
        .on_input(Message::ContentChanged)
        .into()
}

fn update(state: &mut State, message: Message) {
    match message {
        Message::ContentChanged(content) => {
            state.content = content;
        }
    }
}

Implementations

impl<'a, Message, Theme, Renderer> TextInput<'a, Message, Theme, Renderer>
where Message: Clone, Theme: Catalog, Renderer: Renderer,

pub fn new(placeholder: &str, value: &str) -> Self

Creates a new TextInput with the given placeholder and its current value.

pub fn id(self, id: impl Into<Id>) -> Self

Sets the widget::Id of the TextInput.

pub fn secure(self, is_secure: bool) -> Self

Converts the TextInput into a secure password input.

pub fn on_input(self, on_input: impl Fn(String) -> Message + 'a) -> Self

Sets the message that should be produced when some text is typed into the TextInput.

If this method is not called, the TextInput will be disabled.

pub fn on_input_maybe( self, on_input: Option<impl Fn(String) -> Message + 'a>, ) -> Self

Sets the message that should be produced when some text is typed into the TextInput, if Some.

If None, the TextInput will be disabled.

pub fn on_submit(self, message: Message) -> Self

Sets the message that should be produced when the TextInput is focused and the enter key is pressed.

pub fn on_submit_maybe(self, on_submit: Option<Message>) -> Self

Sets the message that should be produced when the TextInput is focused and the enter key is pressed, if Some.

pub fn on_paste(self, on_paste: impl Fn(String) -> Message + 'a) -> Self

Sets the message that should be produced when some text is pasted into the TextInput.

pub fn on_paste_maybe( self, on_paste: Option<impl Fn(String) -> Message + 'a>, ) -> Self

Sets the message that should be produced when some text is pasted into the TextInput, if Some.

pub fn font(self, font: Renderer::Font) -> Self

Sets the Font of the TextInput.

pub fn icon(self, icon: Icon<Renderer::Font>) -> Self

Sets the Icon of the TextInput.

pub fn width(self, width: impl Into<Length>) -> Self

Sets the width of the TextInput.

pub fn padding<P: Into<Padding>>(self, padding: P) -> Self

Sets the [Padding] of the TextInput.

pub fn size(self, size: impl Into<Pixels>) -> Self

Sets the text size of the TextInput.

pub fn line_height(self, line_height: impl Into<LineHeight>) -> Self

Sets the text::LineHeight of the TextInput.

pub fn align_x(self, alignment: impl Into<Horizontal>) -> Self

Sets the horizontal alignment of the TextInput.

pub fn style(self, style: impl Fn(&Theme, Status) -> Style + 'a) -> Self
where Theme::Class<'a>: From<StyleFn<'a, Theme>>,

Sets the style of the TextInput.

pub fn class(self, class: impl Into<Theme::Class<'a>>) -> Self

Sets the style class of the TextInput.

pub fn layout( &mut self, tree: &mut Tree, renderer: &Renderer, limits: &Limits, value: Option<&Value>, ) -> Node

Lays out the TextInput, overriding its Value if provided.

pub fn draw( &self, tree: &Tree, renderer: &mut Renderer, theme: &Theme, layout: Layout<'_>, _cursor: Cursor, value: Option<&Value>, viewport: &Rectangle, )

Draws the TextInput with the given Renderer, overriding its Value if provided.