Struct iced_widget::tooltip::Tooltip

pub struct Tooltip<'a, Message, Renderer = Renderer<Theme>>where
    Renderer: Renderer,
    Renderer::Theme: StyleSheet + StyleSheet,{ /* private fields */ }

An element to display a widget over another.

Implementations

impl<'a, Message, Renderer> Tooltip<'a, Message, Renderer>where Renderer: Renderer, Renderer::Theme: StyleSheet + StyleSheet,

pub fn new( content: impl Into<Element<'a, Message, Renderer>>, tooltip: impl Into<Cow<'a, str>>, position: Position ) -> Self

Creates a new Tooltip.

pub fn size(self, size: impl Into<Pixels>) -> Self

Sets the size of the text of the Tooltip.

pub fn font(self, font: impl Into<Renderer::Font>) -> Self

Sets the font of the Tooltip.

pub fn gap(self, gap: impl Into<Pixels>) -> Self

Sets the gap between the content and its Tooltip.

pub fn padding(self, padding: impl Into<Pixels>) -> Self

Sets the padding of the Tooltip.

pub fn snap_within_viewport(self, snap: bool) -> Self

Sets whether the Tooltip is snapped within the viewport.

pub fn style( self, style: impl Into<<Renderer::Theme as StyleSheet>::Style> ) -> Self

Sets the style of the Tooltip.