Crate iced_widget

Use the built-in widgets or create your own.

Re-exports

pub use button::Button;

pub use checkbox::Checkbox;

pub use column::Column;

pub use combo_box::ComboBox;

pub use container::Container;

pub use list::List;

pub use mouse_area::MouseArea;

pub use pane_grid::PaneGrid;

pub use pick_list::PickList;

pub use progress_bar::ProgressBar;

pub use radio::Radio;

pub use row::Row;

pub use rule::Rule;

pub use scrollable::Scrollable;

pub use slider::Slider;

pub use space::Space;

pub use stack::Stack;

pub use text::Text;

pub use text_editor::TextEditor;

pub use text_input::TextInput;

pub use themer::Themer;

pub use toggler::Toggler;

pub use tooltip::Tooltip;

pub use vertical_slider::VerticalSlider;

pub use svg::Svg;

pub use image::Image;

pub use canvas::Canvas;

pub use iced_renderer as renderer;

pub use iced_renderer::graphics;

pub use iced_runtime as runtime;

pub use iced_runtime::core;

Modules

button

Buttons allow your users to perform actions by pressing them.

canvas

Canvases can be leveraged to draw interactive 2D graphics.

checkbox

Checkboxes can be used to let users make binary choices.

combo_box

Combo boxes display a dropdown list of searchable and selectable options.

container

Containers let you align a widget inside their boundaries.

image

Images display raster graphics in different formats (PNG, JPG, etc.).

keyed

Keyed widgets can provide hints to ensure continuity.

list

overlay

Display interactive elements on top of other widgets.

pane_grid

Pane grids let your users split regions of your application and organize layout dynamically.

pick_list

Pick lists display a dropdown list of selectable options.

progress_bar

Progress bars visualize the progression of an extended computer operation, such as a download, file transfer, or installation.

radio

Radio buttons let users choose a single option from a bunch of options.

rule

Rules divide space horizontally or vertically.

scrollable

Scrollables let users navigate an endless amount of content with a scrollbar.

slider

Sliders let users set a value by moving an indicator.

svg

Svg widgets display vector graphics in your application.

text

Draw and interact with text.

text_editor

Text editors display a multi-line text input for text editing.

text_input

Text inputs display fields that can be filled with text.

theme

Use the built-in theme and styles.

toggler

Togglers let users make binary choices by toggling a switch.

tooltip

Tooltips display a hint of information over some element when hovered.

vertical_slider

Sliders let users set a value by moving an indicator.

Macros

column

Creates a [Column] with the given children.

keyed_column

Creates a keyed Column with the given children.

rich_text

Creates some Rich text with the given spans.

row

Creates a [Row] with the given children.

stack

Creates a Stack with the given children.

text

Creates a new Text widget with the provided content.

Structs

Lazy

A widget that only rebuilds its contents when necessary.

Responsive

A widget that is aware of its dimensions.

Enums

Theme

A built-in theme.

Traits

ComponentDeprecated

A reusable, custom widget that uses The Elm Architecture.

Functions

button

Creates a new Button with the provided content.

canvas

Creates a new Canvas.

center

Creates a new Container that fills all the available space and centers its contents inside.

checkbox

Creates a new Checkbox.

column

Creates a new [Column] with the given children.

combo_box

Creates a new ComboBox.

componentDeprecated

Turns an implementor of Component into an Element that can be embedded in any application.

container

Creates a new Container with the provided content.

focus_next

Focuses the next focusable widget.

focus_previous

Focuses the previous focusable widget.

horizontal_rule

Creates a horizontal Rule with the given height.

horizontal_space

Creates a new [Space] widget that fills the available horizontal space.

hover

Displays a widget on top of another one, only when the base widget is hovered.

iced

Creates an Element that displays the iced logo with the given text_size.

image

Creates a new Image.

keyed_column

Creates a new keyed::Column from an iterator of elements.

lazy

Creates a new Lazy widget with the given data Dependency and a closure that can turn this data into a widget tree.

list

Creates a new List with the provided Content and closure to view an item of the List.

mouse_area

A container intercepting mouse events.

opaque

Wraps the given widget and captures any mouse button presses inside the bounds of the widget—effectively making it opaque.

pane_grid

Creates a PaneGrid with the given pane_grid::State and view function.

pick_list

Creates a new PickList.

progress_bar

Creates a new ProgressBar.

radio

Creates a new Radio.

responsive

Creates a new Responsive widget with a closure that produces its contents.

rich_text

Creates a new Rich text widget with the provided spans.

row

Creates a new [Row] from an iterator.

scrollable

Creates a new Scrollable with the provided content.

slider

Creates a new Slider.

span

Creates a new Span of text with the provided content.

stack

Creates a new Stack with the given children.

svg

Creates a new Svg widget from the given Handle.

text

Creates a new Text widget with the provided content.

text_editor

Creates a new TextEditor.

text_input

Creates a new TextInput.

themer

A widget that applies any Theme to its contents.

toggler

Creates a new Toggler.

tooltip

Creates a new Tooltip for the provided content with the given Element and tooltip::Position.

value

Creates a new Text widget that displays the provided value.

vertical_rule

Creates a vertical Rule with the given width.

vertical_slider

Creates a new VerticalSlider.

vertical_space

Creates a new [Space] widget that fills the available vertical space.

Type Aliases

Renderer

The default graphics renderer for iced.