Crate iced_widget

Use the built-in widgets or create your own.

Re-exports

pub use button::Button;

pub use checkbox::Checkbox;

pub use column::Column;

pub use combo_box::ComboBox;

pub use container::Container;

pub use mouse_area::MouseArea;

pub use pane_grid::PaneGrid;

pub use pick_list::PickList;

pub use progress_bar::ProgressBar;

pub use radio::Radio;

pub use row::Row;

pub use rule::Rule;

pub use scrollable::Scrollable;

pub use slider::Slider;

pub use space::Space;

pub use text::Text;

pub use text_editor::TextEditor;

pub use text_input::TextInput;

pub use themer::Themer;

pub use toggler::Toggler;

pub use tooltip::Tooltip;

pub use vertical_slider::VerticalSlider;

pub use svg::Svg;

pub use image::Image;

pub use canvas::Canvas;

pub use iced_renderer as renderer;

pub use iced_renderer::graphics;

pub use iced_runtime as runtime;

pub use iced_runtime::core;

pub use iced_style as style;

Modules

button

Allow your users to perform actions by pressing a button.

canvas

Draw 2D graphics for your users.

checkbox

Show toggle controls using checkboxes.

combo_box

Display a dropdown list of searchable and selectable options.

container

Decorate content and apply alignment.

image

Display images in your user interface.

keyed

Use widgets that can provide hints to ensure continuity.

overlay

Display interactive elements on top of other widgets.

pane_grid

Let your users split regions of your application and organize layout dynamically.

pick_list

Display a dropdown list of selectable values.

progress_bar

Provide progress feedback to your users.

radio

Create choices using radio buttons.

rule

Display a horizontal or vertical rule for dividing content.

scrollable

Navigate an endless amount of content with a scrollbar.

slider

Display an interactive selector of a single value from a range of values.

space

Distribute content vertically.

svg

Display vector graphics in your application.

text

Draw and interact with text.

text_editor

Display a multi-line text input for text editing.

text_input

Display fields that can be filled with text.

theme

Use the built-in theme and styles.

toggler

Show toggle controls using togglers.

tooltip

Display a widget over another.

vertical_slider

Display an interactive selector of a single value from a range of values.

Macros

column

Creates a Column with the given children.

keyed_column

Creates a Column with the given children.

row

Creates a Row with the given children.

Structs

Lazy

A widget that only rebuilds its contents when necessary.

Responsive

A widget that is aware of its dimensions.

Enums

Renderer

The default graphics renderer for iced.

Theme

A built-in theme.

Traits

Component

A reusable, custom widget that uses The Elm Architecture.

Functions

button

Creates a new Button with the provided content.

canvas

Creates a new Canvas.

checkbox

Creates a new Checkbox.

column

Creates a new [Column] with the given children.

combo_box

Creates a new ComboBox.

component

Turns an implementor of Component into an Element that can be embedded in any application.

container

Creates a new Container with the provided content.

focus_next

Focuses the next focusable widget.

focus_previous

Focuses the previous focusable widget.

horizontal_rule

Creates a horizontal Rule with the given height.

horizontal_space

Creates a new horizontal Space with the given Length.

image

Creates a new Image.

keyed_column

Creates a new keyed::Column with the given children.

lazy

Creates a new Lazy widget with the given data Dependency and a closure that can turn this data into a widget tree.

mouse_area

A container intercepting mouse events.

pick_list

Creates a new PickList.

progress_bar

Creates a new ProgressBar.

radio

Creates a new Radio.

responsive

Creates a new Responsive widget with a closure that produces its contents.

row

Creates a new Row with the given children.

scrollable

Creates a new Scrollable with the provided content.

slider

Creates a new Slider.

svg

Creates a new Svg widget from the given Handle.

text

Creates a new Text widget with the provided content.

text_editor

Creates a new TextEditor.

text_input

Creates a new TextInput.

themer

A widget that applies any Theme to its contents.

toggler

Creates a new Toggler.

tooltip

Creates a new Tooltip with the provided content, tooltip text, and tooltip::Position.

vertical_rule

Creates a vertical Rule with the given width.

vertical_slider

Creates a new VerticalSlider.

vertical_space

Creates a new vertical Space with the given Length.