Module iced_widget::scrollable

Navigate an endless amount of content with a scrollbar.

Structs

AbsoluteOffset

The amount of absolute offset in each direction of a [Scrollable].

Id

The identifier of a Scrollable.

Properties

Properties of a scrollbar within a Scrollable.

RelativeOffset

The amount of relative offset in each direction of a [Scrollable].

Scrollable

A widget that can vertically display an infinite amount of content with a scrollbar.

Scrollbar

The appearance of a scrollable.

Scroller

The appearance of the scroller of a scrollable.

State

The local state of a Scrollable.

Viewport

The current Viewport of the Scrollable.

Enums

Alignment

Alignment of the scrollable’s content relative to it’s Viewport in one direction.

Direction

The direction of Scrollable.

Traits

StyleSheet

A set of rules that dictate the style of a scrollable.

Functions

draw

Draws a Scrollable.

layout

Computes the layout of a Scrollable.

mouse_interaction

Computes the current [mouse::Interaction] of a Scrollable.

scroll_to

Produces a [Command] that scrolls the Scrollable with the given Id to the provided AbsoluteOffset along the x & y axis.

snap_to

Produces a [Command] that snaps the Scrollable with the given Id to the provided percentage along the x & y axis.

update

Processes an [Event] and updates the State of a Scrollable accordingly.