Module iced_widget::vertical_slider

Display an interactive selector of a single value from a range of values.

A VerticalSlider has some local State.

Structs

Appearance

The appearance of a slider.

Handle

The appearance of the handle of a slider.

State

The local state of a VerticalSlider.

VerticalSlider

An vertical bar and a handle that selects a single value from a range of values.

Enums

HandleShape

The shape of the handle of a slider.

Traits

StyleSheet

A set of rules that dictate the style of a slider.

Functions

draw

Draws a VerticalSlider.

mouse_interaction

Computes the current mouse::Interaction of a VerticalSlider.

update

Processes an Event and updates the State of a VerticalSlider accordingly.